refactor: add skeleton in atomic design
