+++ src/atomic/atm.skeleton/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./skeleton.component";
