+++ src/atomic/atm.skeleton/skeleton.component.tsx
@@ -10,4 +10,4 @@ function Skeleton({ className, ...props }: React.ComponentProps<"div">) {
   );
 }
 
-export { Skeleton };
+export default Skeleton;
